Refactor old integration tests to use new integration helpers
Lots of code movement but largely no change in functionality.
Just refactoring to use new test helpers and moving single-use workflow definitions to their corresponding test file usages.
