--- packages/test/src/test-integration-old.ts
@@ -1,12 +1,0 @@
-/**
- * This file has been given the suffix -old because it uses an older style of
- * integration testing. New code should follow the style of integration tests in
- * the files without this suffix.
- */
-
-import { RUN_INTEGRATION_TESTS } from './helpers';
-import { runIntegrationTests } from './integration-tests-old';
-
-if (RUN_INTEGRATION_TESTS) {
-  runIntegrationTests();
-}
