--- packages/test/src/workflows/child-workflow-failure.ts
@@ -1,13 +1,0 @@
-/**
- * Tests child workflow failure from the parent workflow perspective
- * @module
- */
-
-import { executeChild } from '@temporalio/workflow';
-import { throwAsync } from './throw-async';
-
-export async function childWorkflowFailure(): Promise<void> {
-  await executeChild(throwAsync, {
-    taskQueue: 'test',
-  });
-}
