--- packages/test/src/workflows/child-workflow-timeout.ts
@@ -1,14 +1,0 @@
-/**
- * Tests child workflow timeout from the parent workflow perspective
- * @module
- */
-
-import { executeChild } from '@temporalio/workflow';
-import { unblockOrCancel } from './unblock-or-cancel';
-
-export async function childWorkflowTimeout(): Promise<void> {
-  await executeChild(unblockOrCancel, {
-    workflowExecutionTimeout: '10ms',
-    retry: { maximumAttempts: 1 },
-  });
-}
