+++ packages/test/src/workflows/index.ts
@@ -1,5 +1,5 @@
+// TODO(thomas): use filenames to organize into workflows
 export * from './abort-controller';
-export * from './activity-failure';
 export * from './activity-failures';
 export * from './args-and-return';
 export * from './async-activity-completion-tester';
@@ -12,19 +12,13 @@ export * from './cancel-requested-with-non-cancellable';
 export * from './cancel-scope-on-failed-validation';
 export * from './cancel-timer-immediately';
 export * from './cancel-timer-immediately-alternative-impl';
-export * from './cancel-timer-with-delay';
 export * from './cancel-workflow';
 export * from './cancellation-error-is-propagated';
 export * from './cancellation-scopes';
 export * from './cancellation-scopes-with-callbacks';
 export * from './child-and-noncancellable';
 export * from './child-workflow-cancel';
-export * from './child-workflow-failure';
-export * from './child-workflow-invoke';
 export * from './child-workflow-signals';
-export * from './child-workflow-start-fail';
-export * from './child-workflow-termination';
-export * from './child-workflow-timeout';
 export * from './condition';
 export * from './condition-completion-race';
 export * from './condition-timeout-0';
@@ -60,12 +54,9 @@ export * from './patched-top-level';
 export * from './promise-all';
 export * from './promise-race';
 export * from './promise-then-promise';
-export * from './query-and-condition';
 export * from './race';
 export * from './random';
 export * from './reject-promise';
-export * from './return-search-attributes';
-export * from './return-workflow-info';
 export * from './run-activity-in-different-task-queue';
 export * from './scope-cancelled-while-waiting-on-external-workflow-cancellation';
 export * from './set-timeout-after-microtasks';
@@ -80,7 +71,6 @@ export * from './signal-update-ordering';
 export * from './signals-timers-activities-order';
 export * from './sinks';
 export * from './sleep';
-export * from './sleep-invalid-duration';
 export * from './smorgasbord';
 export * from './stack-tracer';
 export * from './success-string';
@@ -88,8 +78,6 @@ export * from './swc';
 export * from './tasks-and-microtasks';
 export * from './text-encoder-decoder';
 export * from './throw-async';
-export * from './throw-big-int';
-export * from './throw-object';
 export * from './trailing-timer';
 export * from './try-to-continue-after-completion';
 export * from './two-strings';
@@ -98,7 +86,6 @@ export * from './two-strings';
 export { isBlockedQuery, unblockOrCancel } from './unblock-or-cancel';
 export * from './unhandled-rejection';
 export * from './upsert-and-read-search-attributes';
-export * from './url-whatwg';
 export * from './wait-on-user';
 export * from './workflow-cancellation-scenarios';
 export * from './upsert-and-read-memo';
