--- packages/test/src/workflows/return-workflow-info.ts
@@ -1,5 +1,0 @@
-import { WorkflowInfo, workflowInfo } from '@temporalio/workflow';
-
-export async function returnWorkflowInfo(): Promise<WorkflowInfo> {
-  return workflowInfo();
-}
