--- packages/test/src/workflows/throw-big-int.ts
@@ -1,3 +1,0 @@
-export async function throwBigInt(): Promise<void> {
-  throw 42n;
-}
