--- packages/test/src/workflows/throw-object.ts
@@ -1,3 +1,0 @@
-export async function throwObject(): Promise<void> {
-  throw { plainObject: true };
-}
